1.3

+++ src/main/java/web/dao/UserDAOImpl.java
@@ -2,15 +2,12 @@ package web.dao;
 
 
 import org.springframework.stereotype.Repository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import web.model.User;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
-//@Service
 @Repository
 public class UserDAOImpl implements UserDAO {
     @PersistenceContext
@@ -22,25 +19,22 @@ public class UserDAOImpl implements UserDAO {
     }
 
     @Override
-    public User getUser(int id) {
+    public User getUser(Long id) {
         return entityManager.find(User.class, id);
     }
 
     @Override
-    //@Transactional
     public void save(User user) {
         entityManager.merge(user);
     }
 
     @Override
-    //@Transactional
-    public void delete(int id) {
+    public void delete(Long id) {
         User user = getUser(id);
         entityManager.remove(user);
     }
 
     @Override
-    //@Transactional
     public void update(User user) {
         entityManager.merge(user);
     }
