+++ src/main/java/web/model/User.java
@@ -8,7 +8,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -25,7 +25,7 @@ public class User {
     public User() {
     }
 
-    public User(int id, String name, String surname, int age, String email) {
+    public User(Long id, String name, String surname, int age, String email) {
         this.id = id;
         this.name = name;
         this.surname = surname;
@@ -33,11 +33,11 @@ public class User {
         this.email = email;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
